Changing re-find to re-matches in regex predicate
Using re-find returns the first match of the regex, re-matches checks
that the entire string matches the regex.

Re-find is unsuitable here as it allows for trailing characters in the
string that aren't alloed by the regex. This was highlighted when
working with email strings. Email strings with numerals in the tld
where allowed by re-find when at the end of the string. The regex
explicitly states that no numerals should be in the tld.

Email regexp has been change to include a $ at the end to ensure it is
complete, and the email portion separated out for reuse.